Add: size_t for memory size and indexing of spsc queue

+++ src/utils/queues/SpscQueue.hpp
@@ -9,7 +9,8 @@ namespace anasa
 
 // One producer and one consumer only.
 // One slot stays empty so full and empty are easy to distinguish.
-template <class T, int capacity>
+// capacity, _read and _write are size_t: they can reach max possible object size in memory 
+template <class T, size_t capacity>
 class SpscQueue
 {
     static_assert(capacity >= 2, "SpscQueue capacity must be at least 2");
@@ -17,8 +18,8 @@ class SpscQueue
     public:
         bool push(const T& item)
         {
-            int write = _write.load(std::memory_order_relaxed);
-            int next = (write + 1) % capacity;
+            size_t write = _write.load(std::memory_order_relaxed);
+            size_t next = (write + 1) % capacity;
 
             if (next == _read.load(std::memory_order_acquire))
                 return false;
@@ -30,7 +31,7 @@ class SpscQueue
 
         bool pop(T& item)
         {
-            int read = _read.load(std::memory_order_relaxed);
+            size_t read = _read.load(std::memory_order_relaxed);
 
             if (read == _write.load(std::memory_order_acquire))
                 return false;
@@ -43,7 +44,7 @@ class SpscQueue
         bool peek(T& item) const
         {
             /* Returns the front element but leaves it in the queue */
-            int read = _read.load(std::memory_order_relaxed);
+            size_t read = _read.load(std::memory_order_relaxed);
 
             if (read == _write.load(std::memory_order_acquire))
                 return false;
@@ -52,21 +53,21 @@ class SpscQueue
             return true;
         }
 
-        int usableCapacity() const
+        size_t usableCapacity() const
         {
             return capacity - 1;
         }
 
         bool isEmpty() const
         {
-            int read = _read.load(std::memory_order_relaxed);
+            size_t read = _read.load(std::memory_order_relaxed);
             return read == _write.load(std::memory_order_acquire);
         }
 
     private:
         std::array<T, capacity> _data{};
-        std::atomic<int> _write{0};
-        std::atomic<int> _read{0};
+        std::atomic<size_t> _write{0};
+        std::atomic<size_t> _read{0};
 };
 
 } // namespace anasa
